Calling function to TransferMatrix
This is the file which can call one of TransferMatrix_multiple or TransferMatrix_packing to evaluate results corresponding to added details.
User can manipulate the wavelength range and the thickness of the TMM model, the names for n,k have to be evaluated from the Index_of_Refraction_library.xls. If not already mentioned, new n,k data in the specified range of 300-800nm have to be added to the xls file using addNK_to_TransferMatrix.
Multiple comma-separted .txt reflectance file can be stored.