The illuminant file
The code, which is basically an adapted version of the matlab code from https://web.stanford.edu/group/mcgehee/transfermatrix/, uses the AM1.5G solar spectrum for calculation of few results, which is irrelevant to our usecase. However, for the successful running of the code, we need this file to be present in the main folder.