The main function for TMM simulation
This function deploys the TMM software as mentioned in https://web.stanford.edu/group/mcgehee/transfermatrix/. It has been modified to include the Effective Medium Approximation. This function is called from the multiple_reflectance file and can save reflectance .txt for system with variable thickness of certain layer, which can be modified - the number of layer and the step size of variation.