Another function for TMM simulation
This is another main function which deploys the TMM package originally created by https://web.stanford.edu/group/mcgehee/transfermatrix/. It also includes the Effective Medium Approximation. This has another purpose though. It simulates packing, in the sense that it should be utilized when EMA has to be definitely employed, otherwise it is no different than TransferMatrix_multiple. 
This file save reflectance .txt for variable packing fraction of the layer, which utilizes TMM, for a constant thickness. Unlike the other main function file, which calculates reflectance for multiple thickness for a fixed packing fraction.