Makes colorbar, usable with TransferMatrix_packing
This is a python file which plots a colorbar, needing a user input to select the folder containing the file, filename format, and maximum thickness to search for in the filenames. Use this file when plotting colorbar for reflectance files created using TransferMatrix_packing, since it plots for constant thickness and variable packing fraction.