Generates a heatmap from the CIEUCS colorspace
This file generated a TCD map, which basically calculates the euclidean distance of colors from the CIEUCS colorspace and arrange it in ordered format by assigning a colormap which can be chnaged by the user. This also saves the generated TCD in normalized fashion with the corresponding thickness in a .txt file which needs to be used in the next file, which uses this heatmap as a template to find thickness from euclidean distance.