The main file
The user can select an image which is converted from sRGB to CIEUCS space and then to the euclidean map or the TCD map. The reference heatmap generated in the previous step is loaded which the correlates the TCD range to thickness. The values of TCD have been distributed equally in compartments whose size can be chosen by the user.